pytest_axe/tests/requirements/tests.txt
pytest==3.8.2
pytest-selenium==1.14.0
axe_selenium_python==2.1.1
